+++ include/scaler/image_base.hh
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <algorithm>
 #include <scaler/compiler_compat.hh>
 #include <scaler/vec3.hh>
 namespace scaler {
@@ -10,26 +9,26 @@ namespace scaler {
     template<typename Derived, typename PixelType = uvec3>
     class InputImageBase {
         public:
-            [[nodiscard]] inline int width() const noexcept {
+            [[nodiscard]] inline size_t width() const noexcept {
                 return static_cast<const Derived*>(this)->width_impl();
             }
 
-            [[nodiscard]] inline int height() const noexcept {
+            [[nodiscard]] inline size_t height() const noexcept {
                 return static_cast<const Derived*>(this)->height_impl();
             }
 
-            [[nodiscard]] inline PixelType get_pixel(int x, int y) const noexcept {
+            [[nodiscard]] inline PixelType get_pixel(size_t x, size_t y) const noexcept {
                 return static_cast<const Derived*>(this)->get_pixel_impl(x, y);
             }
 
             [[nodiscard]] inline PixelType safeAccess(int x, int y, OutOfBoundsStrategy strategy = NEAREST) const noexcept {
                 // Optimize for the common case where access is within bounds
                 // Use likely/unlikely hints for branch prediction
-                const int w = width();
-                const int h = height();
+                const size_t w = width();
+                const size_t h = height();
 
-                if (SCALER_LIKELY((x >= 0 && x < w) && (y >= 0 && y < h))) {
-                    return get_pixel(x, y);
+                if (SCALER_LIKELY((x >= 0 && static_cast<size_t>(x) < w) && (y >= 0 && static_cast<size_t>(y) < h))) {
+                    return get_pixel(static_cast<size_t>(x), static_cast<size_t>(y));
                 }
 
                 // Handle out-of-bounds cases
@@ -37,9 +36,9 @@ namespace scaler {
                     case ZERO:
                         return {};
                     case NEAREST:
-                        x = (x < 0) ? 0 : (x >= w) ? (w - 1) : x;
-                        y = (y < 0) ? 0 : (y >= h) ? (h - 1) : y;
-                        return get_pixel(x, y);
+                        x = (x < 0) ? 0 : (static_cast<size_t>(x) >= w) ? (static_cast<int>(w) - 1) : x;
+                        y = (y < 0) ? 0 : (static_cast<size_t>(y) >= h) ? (static_cast<int>(h) - 1) : y;
+                        return get_pixel(static_cast<size_t>(x), static_cast<size_t>(y));
                 }
 
                 return {};  // Unreachable, but keeps compiler happy
@@ -50,15 +49,15 @@ namespace scaler {
     template<typename Derived, typename PixelType = uvec3>
     class OutputImageBase {
         public:
-            [[nodiscard]] int width() const {
+            [[nodiscard]] size_t width() const {
                 return static_cast<const Derived*>(this)->width_impl();
             }
 
-            [[nodiscard]] int height() const {
+            [[nodiscard]] size_t height() const {
                 return static_cast<const Derived*>(this)->height_impl();
             }
 
-            void set_pixel(int x, int y, const PixelType& pixel) {
+            void set_pixel(size_t x, size_t y, const PixelType& pixel) {
                 static_cast<Derived*>(this)->set_pixel_impl(x, y, pixel);
             }
