+++ include/scaler/xbr.hh
@@ -41,7 +41,7 @@ namespace scaler {
 
     // Generic XBR scaler using CRTP - works with any image implementation
     template<typename InputImage, typename OutputImage>
-    auto scaleXbr(const InputImage& src, int scale_factor = 2)
+    auto scaleXbr(const InputImage& src, size_t scale_factor = 2)
         -> OutputImage {
         OutputImage result(src.width() * scale_factor, src.height() * scale_factor, src);
 
@@ -50,13 +50,13 @@ namespace scaler {
         SlidingWindow5x5 <PixelType> window(src.width());
         window.initialize(src, 0);
 
-        for (int y = 0; y < src.height(); y++) {
+        for (size_t y = 0; y < src.height(); y++) {
             // Advance sliding window for next row
             if (y > 0) {
                 window.advance(src);
             }
 
-            for (int x = 0; x < src.width(); x++) {
+            for (size_t x = 0; x < src.width(); x++) {
                 // Get 5x5 neighborhood from cache-friendly buffer
                 PixelType neighborhood[5][5];
                 window.getNeighborhood(x, neighborhood);
@@ -204,8 +204,8 @@ namespace scaler {
                     }
                 }
 
-                int dst_x = scale_factor * x;
-                int dst_y = scale_factor * y;
+                size_t dst_x = scale_factor * x;
+                size_t dst_y = scale_factor * y;
                 result.set_pixel(dst_x, dst_y, top_left_pixel);
                 result.set_pixel(dst_x + 1, dst_y, top_right_pixel);
                 result.set_pixel(dst_x, dst_y + 1, bot_left_pixel);
